moves start func option to test file, adds slow start test to own file that doesnt build when testing with race flag
